classifier/pyproject.toml
[tool.poetry]
name = "classifier"
version = "0.1.0"
description = "Classifies Proposals from RabbitMQ and saves results to the database"
authors = ["SolDevelo <sales@soldevelo.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
openrouter-haystack = "^0.1.0"
haystack-ai = "^2.0.0"
tiktoken = "^0.8.0"
sentence-transformers = "^3.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
